Append a hostname to the directory name for saving
When the saving directory is shared among separate hosts and multiple
sheetreaders run at the same time, there is a case that the directory
name that sheetreader generates for saving is a same name. It leads
to losing data. In order to keep the uniqueness of the directory name,
append a hostname to the directory name.